Use structuredClone in deepClone with fallback
