+++ src/utils/cloneUtils.js
@@ -46,6 +46,10 @@ export function deepClone(value) {
     return value;
   }
 
+  if (typeof structuredClone === 'function') {
+    return structuredClone(value);
+  }
+
   return JSON.parse(JSON.stringify(value));
 }
